@@ -1,15 +1,16 @@
 {
   "name": "@aloskutov/eleventy-plugin-codepen",
-  "version": "1.2.8",
+  "version": "2.0.0",
   "description": "Eleventy codepen plugin",
+  "type": "module",
   "publishConfig": {
     "access": "public"
   },
   "main": ".eleventy.js",
   "homepage": "https://github.com/aloskutov/eleventy-plugin-codepen",
   "scripts": {
-    "test": "jest",
-    "coverage": "jest --coverage",
+    "test": " node --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "coverage": " node --experimental-vm-modules node_modules/jest/bin/jest.js --coverage",
     "lint": "npx eslint ./src",
     "jshint": "npx jshint ./src",
     "demo": "npx @11ty/eleventy --serve"
@@ -33,9 +34,13 @@
     "url": "https://github.com/aloskutov/eleventy-plugin-codepen/issues"
   },
   "devDependencies": {
-    "eslint": "^8.56.0",
+    "@eslint/eslintrc": "^3.3.3",
+    "@eslint/js": "^9.39.2",
+    "babel-jest": "^30.2.0",
+    "eslint": "^9.39.2",
     "eslint-config-google": "^0.14.0",
-    "jest": "^29.7.0",
+    "globals": "^17.2.0",
+    "jest": "^30.2.0",
     "jshint": "^2.13.6"
   }
 }
